+++ src/controllers/CampaignTypesController.php
@@ -32,14 +32,13 @@ class CampaignTypesController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require permission
         $this->requirePermission('campaign:settings');
+
+        return parent::beforeAction($action);
     }
 
     /**