+++ src/controllers/CampaignsController.php
@@ -39,14 +39,13 @@ class CampaignsController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require permission
         $this->requirePermission('campaign:campaigns');
+
+        return parent::beforeAction($action);
     }
 
     /**