+++ src/controllers/ImportsController.php
@@ -37,14 +37,13 @@ class ImportsController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require permission
         $this->requirePermission('campaign:importContacts');
+
+        return parent::beforeAction($action);
     }
 
     /**