+++ src/controllers/MailingListTypesController.php
@@ -31,14 +31,13 @@ class MailingListTypesController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require permission
         $this->requirePermission('campaign:settings');
+
+        return parent::beforeAction($action);
     }
 
     /**