+++ src/controllers/MailingListsController.php
@@ -33,14 +33,13 @@ class MailingListsController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require permission
         $this->requirePermission('campaign:mailingLists');
+
+        return parent::beforeAction($action);
     }
 
     /**