+++ src/controllers/SegmentsController.php
@@ -35,17 +35,16 @@ class SegmentsController extends Controller
 
     /**
      * @inheritdoc
-     * @throws ForbiddenHttpException
      */
-    public function init()
+    public function beforeAction($action): bool
     {
-        parent::init();
-
         // Require pro
         Campaign::$plugin->requirePro();
 
         // Require permission
         $this->requirePermission('campaign:segments');
+
+        return parent::beforeAction($action);
     }
 
     /**