+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - The `@web` alias warning is now only displayed when not explicitly defined ([#208](https://github.com/putyourlightson/craft-campaign/issues/208)).
 - Improved formatting of UI elements.
+- Improved enforcing of permissions in controllers.
 
 ## 1.18.0 - 2021-01-04
 > {tip} Read the updated docs on [Sending to Large Lists](https://putyourlightson.com/plugins/campaign#sending-to-large-lists).